aggiunta della barra della vita per il player e refactoring per un nuovo
file fxml per gestire l'hud del player

+++ src/main/java/it/unicam/cs/mpgc/rpg126598/model/Targetable.java
@@ -1,8 +1,8 @@
 package it.unicam.cs.mpgc.rpg126598.model;
 
 public interface Targetable {
-    void takeDamage(int amount);
+    void takeDamage(double amount);
     boolean isDead();
-    int getDefense();
+    double getDefense();
 
 }
